Support authentication through client SSL certificates
Add support for authentication through client ssl certificates.
In our use case, we use Belgian eID's (smartcards) to unlock the certificate,
but this should not matter (much).
Implementation is done using a separate auth_type and login_eid.php file,
Authentication is done by taking the serial number in the
SSL_CLIENT_S_DN field, sprintf()ing this in a configurable string to
construct a username DN, and combining it with a global - configurable -
password to do a normal ldap user login.
To use this, your webserver must be set up properly (in the case of apache:
use mod_ssl, and use the appropriate SSLCACertificateFile)
This should integrate nicely and should not interfere with the existing
authentication system.